# Speed up folds on Sequences