… benchmarks/Sequence.hs
… containers.cabal
… tests/seq-properties.hs