As referenced in  #504.

After noticing that `foldMapWithIndex` was significantly faster than `foldMap`, I rewrote the `Foldable` methods to mimic the style of `foldMapWithIndex`. Writing the first level of recursion out on the finger tree manually, and specialising the folds on nodes and digits manually, yields a significant speedup. From my testing, there's a ~3.8x speedup for `foldMap`, a ~1.8x speedup for `foldl'` and `foldr'`, and an ~8x speedup for `foldl` and `foldr`.

These are the functions benchmarked:

```haskell
foldMapSum :: Seq.Seq Int -> Int
foldMapSum = getSum . foldMap Sum

foldlSum :: Seq.Seq Int -> Int
foldlSum xs = foldl (\k x z -> k $! z+x) id xs 0

foldlSum' :: Seq.Seq Int -> Int
foldlSum' = foldl' (+) 0

foldrSum :: Seq.Seq Int -> Int
foldrSum xs = foldr (\x k z -> k $! z+x) id xs 0

foldrSum' :: Seq.Seq Int -> Int
foldrSum' = foldr' (+) 0
```

And these are the results, when run on random sequences of length 500000:

```
benchmarking 500000/foldMapSum/new
time                 3.483 ms   (3.457 ms .. 3.504 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 3.429 ms   (3.398 ms .. 3.458 ms)
std dev              100.7 μs   (79.44 μs .. 153.5 μs)
variance introduced by outliers: 13% (moderately inflated)

benchmarking 500000/foldMapSum/old
time                 13.30 ms   (12.62 ms .. 14.20 ms)
                     0.981 R²   (0.968 R² .. 0.992 R²)
mean                 12.41 ms   (12.12 ms .. 12.85 ms)
std dev              916.9 μs   (674.8 μs .. 1.243 ms)
variance introduced by outliers: 37% (moderately inflated)

benchmarking 500000/foldrSum/new
time                 4.349 ms   (4.274 ms .. 4.437 ms)
                     0.998 R²   (0.997 R² .. 0.999 R²)
mean                 3.991 ms   (3.916 ms .. 4.068 ms)
std dev              197.3 μs   (166.6 μs .. 255.0 μs)
variance introduced by outliers: 27% (moderately inflated)

benchmarking 500000/foldrSum/old
time                 37.39 ms   (36.83 ms .. 38.49 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 35.95 ms   (35.35 ms .. 36.56 ms)
std dev              1.151 ms   (598.5 μs .. 1.561 ms)

benchmarking 500000/foldrSum'/new
time                 2.725 ms   (2.695 ms .. 2.774 ms)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 2.694 ms   (2.665 ms .. 2.728 ms)
std dev              103.0 μs   (78.28 μs .. 143.5 μs)
variance introduced by outliers: 21% (moderately inflated)

benchmarking 500000/foldrSum'/old
time                 4.926 ms   (4.809 ms .. 5.116 ms)
                     0.997 R²   (0.994 R² .. 0.999 R²)
mean                 4.811 ms   (4.762 ms .. 4.883 ms)
std dev              166.3 μs   (124.8 μs .. 211.4 μs)
variance introduced by outliers: 15% (moderately inflated)

benchmarking 500000/foldlSum/new
time                 5.207 ms   (5.071 ms .. 5.310 ms)
                     0.996 R²   (0.994 R² .. 0.998 R²)
mean                 4.672 ms   (4.571 ms .. 4.803 ms)
std dev              315.6 μs   (263.3 μs .. 363.5 μs)
variance introduced by outliers: 40% (moderately inflated)

benchmarking 500000/foldlSum/old
time                 38.99 ms   (38.09 ms .. 39.73 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 37.77 ms   (36.82 ms .. 38.26 ms)
std dev              1.307 ms   (622.3 μs .. 1.860 ms)

benchmarking 500000/foldlSum'/new
time                 2.452 ms   (2.413 ms .. 2.487 ms)
                     0.999 R²   (0.998 R² .. 0.999 R²)
mean                 2.359 ms   (2.337 ms .. 2.379 ms)
std dev              64.66 μs   (54.28 μs .. 77.18 μs)
variance introduced by outliers: 13% (moderately inflated)

benchmarking 500000/foldlSum'/old
time                 4.418 ms   (4.273 ms .. 4.548 ms)
                     0.996 R²   (0.994 R² .. 0.999 R²)
mean                 4.188 ms   (4.147 ms .. 4.244 ms)
std dev              139.5 μs   (102.4 μs .. 182.2 μs)
variance introduced by outliers: 16% (moderately inflated)
```

For reference, the `foldMapWithIndex` function could be used to sum as well:

```haskell
foldMapWithIndexSum :: Seq.Seq Int -> Int
foldMapWithIndexSum = getSum . Seq.foldMapWithIndex (const Sum)
```

And these are its times (unchanged between old/new):

```
benchmarking 500000/foldMapWithIndexSum
time                 4.476 ms   (4.407 ms .. 4.532 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 4.351 ms   (4.312 ms .. 4.404 ms)
std dev              129.5 μs   (106.3 μs .. 183.3 μs)
variance introduced by outliers: 12% (moderately inflated)
```