@oisdk 2018-01-23
(oh, and I noticed a typo I had added in the sequence benchmark file)