@treeowl 2018-01-23
Just a couple more things to check:

1. How are the foldl and foldr default definitions using your new foldMap?
2. How are the foldl' and foldr' default definitions using your new foldr
and foldl?

I don't imagine we'll be able to use the defaults, but it's worth a shot to
keep the already-enormous source code size down.

On Tue, Jan 23, 2018 at 3:10 PM, Donnacha Oisín Kidney <
notifications@github.com> wrote:

> (oh, and I noticed a typo I had added in the sequence benchmark file)
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/haskell/containers/pull/510#issuecomment-359914966>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ABzi_ducqfSjOOFE0GK7lOGEYNjuZQtiks5tNjzSgaJpZM4RqPjy>
> .
>