@oisdk 2018-01-23
`foldl` and `foldr` go back to their original speeds (pretty much exactly) on the default definition, and the default `foldr'` and `foldl'` are actually slower than what was there before.