@treeowl 2018-01-23
Any sense of why the old versions of `foldr` and `foldl` were better for implementing `foldl'` and `foldr'` than the new and improved ones? How much better were they?