@treeowl 2018-01-23
I'm not sure your `foldl` and `foldr` benchmarks are really the best ones to use. Sums should surely use strict folds. `foldl` and `foldr` are mainly for

1. Converting to lists or list-like structures and/or
2. Folding with one or more strict accumulators.

I want things like this to be efficient, if possible:

```haskell
f x0 y0 z0 as= foldr go stop as x0 y0 z0 where
  go a r !x !y !z = ...
  stop !x !y !z = ...

g x0 as = foldr go stop as x0 where
  go a r !x = let !(a',x') = p a x in a' : r x'
```