@oisdk 2018-01-23
The old `foldl'` and `foldr'` were written manually, whereas the default implementations called in to `foldr` and `foldl`. I haven't benchmarked what the default implementations which rely on the *old* `foldl` and `foldr` would do, but I'd imagine it's similar to the current measurements for `foldlSum` and `foldrSum`, as those functions I pretty much lifted from the default definitions of `foldl`  and `foldr`. (the default `foldl'` and `foldr'` were ~15ms)