@treeowl 2018-01-23
Oh, sorry, my memory was out of date. We used to use the defaults for `foldl'` and `foldr'`.