@treeowl 2018-01-23
FYI: two mashed places that come to mind are `splitMap` (used to implement `zipWith` and `chunksOf`) and the `Applicative` machinery. I don't know if there's a way to modify `aptyMiddle` or any of the rest of that to make it sensible in any other context, but it might be worth thinking about.