@oisdk 2018-01-23
I'll look for a good candidate for multiple strict accumulators for `foldr`, but in the meantime `toList` has the following improvement:

```
benchmarking 500000/toList/new
time                 14.10 ms   (13.87 ms .. 14.34 ms)
                     0.998 R²   (0.995 R² .. 0.999 R²)
mean                 13.67 ms   (13.51 ms .. 13.86 ms)
std dev              456.7 μs   (346.3 μs .. 629.7 μs)
variance introduced by outliers: 11% (moderately inflated)
             
benchmarking 500000/toList/old
time                 18.96 ms   (18.56 ms .. 19.35 ms)
                     0.997 R²   (0.995 R² .. 0.999 R²)
mean                 18.17 ms   (17.73 ms .. 18.51 ms)
std dev              902.7 μs   (499.3 μs .. 1.552 ms)
variance introduced by outliers: 18% (moderately inflated)
```