@treeowl 2018-01-23
GHC 7.8 needs a different implementation of `(.#)` or maybe a different type signature for it. Check what `Data.Profunctors.Unsafe` does.