@treeowl 2018-01-23
Actually, we don't need to be as fancy as `Data.Profunctors.Unsafe`, because we know what types we're coercing. Just swap the arguments to `Coercible` in the type signature.