@oisdk 2018-01-23
So I've taken some of the "FB" forms of functions from Data.List:

```haskell
foldrTake :: Int -> Seq.Seq Int -> [Int]
foldrTake n xs = foldr (\x xs m -> case m of 1 -> [x]; _ -> x : xs (m-1)) (const []) xs n

foldrScanl :: Seq.Seq Int -> [Int]
foldrScanl bs = 0 : foldr (\b g -> oneShot (\x -> let !b' = x + b in b' : g b')) (const []) bs 0
```

And here are the results:

```
benchmarking 500000/foldrScanl/new
time                 13.35 ms   (12.04 ms .. 14.71 ms)
                     0.963 R²   (0.933 R² .. 0.987 R²)
mean                 14.09 ms   (13.62 ms .. 14.76 ms)
std dev              1.413 ms   (1.070 ms .. 1.841 ms)
variance introduced by outliers: 48% (moderately inflated)
             
benchmarking 500000/foldrScanl/old
time                 55.00 ms   (52.27 ms .. 58.46 ms)
                     0.993 R²   (0.985 R² .. 0.997 R²)
mean                 47.31 ms   (44.97 ms .. 49.78 ms)
std dev              4.565 ms   (3.825 ms .. 5.301 ms)
variance introduced by outliers: 37% (moderately inflated)
             
benchmarking 500000/foldrTake/new
time                 6.626 ms   (6.199 ms .. 7.081 ms)
                     0.952 R²   (0.911 R² .. 0.980 R²)
mean                 6.061 ms   (5.839 ms .. 6.432 ms)
std dev              871.2 μs   (629.0 μs .. 1.357 ms)
variance introduced by outliers: 73% (severely inflated)
             
benchmarking 500000/foldrTake/old
time                 23.09 ms   (21.65 ms .. 25.00 ms)
                     0.981 R²   (0.960 R² .. 0.997 R²)
mean                 22.55 ms   (21.52 ms .. 23.37 ms)
std dev              1.999 ms   (1.438 ms .. 2.884 ms)
variance introduced by outliers: 35% (moderately inflated)
```